#1218 - organize files better
