+++ src/frontend/src/pages/AdminToolsPage/FinanceConfig/CreateAccountCodeModal.tsx
@@ -1,6 +1,6 @@
-import LoadingIndicator from '../../components/LoadingIndicator';
-import { useCreateAccountCode } from '../../hooks/finance.hooks';
-import ErrorPage from '../ErrorPage';
+import LoadingIndicator from '../../../components/LoadingIndicator';
+import { useCreateAccountCode } from '../../../hooks/finance.hooks';
+import ErrorPage from '../../ErrorPage';
 import AccountCodeFormModal from './AccountCodeFormModal';
 
 interface CreateAccountCodeModalProps {
