+++ src/frontend/src/pages/AdminToolsPage/FinanceConfig/CreateVendorModal.tsx
@@ -1,11 +1,11 @@
 import { useForm } from 'react-hook-form';
-import NERFormModal from '../../components/NERFormModal';
+import NERFormModal from '../../../components/NERFormModal';
 import { FormControl, FormLabel } from '@mui/material';
-import ReactHookTextField from '../../components/ReactHookTextField';
-import { useCreateVendor } from '../../hooks/finance.hooks';
-import { useToast } from '../../hooks/toasts.hooks';
-import ErrorPage from '../ErrorPage';
-import LoadingIndicator from '../../components/LoadingIndicator';
+import ReactHookTextField from '../../../components/ReactHookTextField';
+import { useCreateVendor } from '../../../hooks/finance.hooks';
+import { useToast } from '../../../hooks/toasts.hooks';
+import ErrorPage from '../../ErrorPage';
+import LoadingIndicator from '../../../components/LoadingIndicator';
 import * as yup from 'yup';
 import { yupResolver } from '@hookform/resolvers/yup';
