+++ src/frontend/src/pages/AdminToolsPage/FinanceConfig/VendorsTable.tsx
@@ -1,9 +1,9 @@
 import { TableRow, TableCell, Paper, Table, TableBody, TableContainer, TableHead, Typography, Box } from '@mui/material';
-import LoadingIndicator from '../../components/LoadingIndicator';
-import { useGetAllVendors } from '../../hooks/finance.hooks';
-import { datePipe } from '../../utils/pipes';
-import ErrorPage from '../ErrorPage';
-import { NERButton } from '../../components/NERButton';
+import LoadingIndicator from '../../../components/LoadingIndicator';
+import { useGetAllVendors } from '../../../hooks/finance.hooks';
+import { datePipe } from '../../../utils/pipes';
+import ErrorPage from '../../ErrorPage';
+import { NERButton } from '../../../components/NERButton';
 import { useState } from 'react';
 import CreateVendorModal from './CreateVendorModal';
